+++ packages/rspack-test-tools/src/case/normal.ts
@@ -1,19 +1,20 @@
 import path from "path";
 import { RspackNormalProcessor } from "../processor/normal";
 import { BasicCaseCreator } from "../test/creator";
+import { NormalRunnerFactory } from "../runner";
 
 const creator = new BasicCaseCreator({
 	clean: true,
-	runable: true,
 	describe: true,
-	steps: ({ name }, testConfig) => [
+	steps: ({ name }) => [
 		new RspackNormalProcessor({
 			name,
 			root: path.resolve(__dirname, "../../../rspack"),
 			compilerOptions: {}, // do not used in rspack
-			testConfig
+			runable: true
 		})
-	]
+	],
+	runner: NormalRunnerFactory
 });
 
 export function createNormalCase(name: string, src: string, dist: string) {
