+++ packages/rspack-test-tools/src/runner/index.ts
@@ -1,7 +1,6 @@
 export * from "./basic";
-export * from "./esm";
-export * from "./web";
 export * from "./hot";
-export * from "./watch";
 export * from "./normal";
+export * from "./watch";
 export * from "./type";
+export * from "./multiple";
