+++ packages/rspack-test-tools/src/test/tester.ts
@@ -6,9 +6,6 @@ import {
 	ITestProcessor,
 	ITestEnv
 } from "../type";
-import createLazyTestEnv from "../helper/legacy/createLazyTestEnv";
-import path from "path";
-import fs from "fs";
 
 export class Tester implements ITester {
 	private context: ITestContext;
@@ -123,22 +120,5 @@ export class Tester implements ITester {
 				console.error(error);
 			}
 		}
-	}
-
-	static createLazyTestEnv(): ITestEnv {
-		return createLazyTestEnv(10000);
-	}
-
-	static createTestEnv(): ITestEnv {
-		return {
-			it,
-			beforeEach,
-			afterEach
-		};
-	}
-
-	static isSkipped(config: { name: string; casePath: string }) {
-		const filterPath = path.join(config.casePath, "test.filter.js");
-		return fs.existsSync(filterPath) && !require(filterPath)(config);
 	}
 }
